Add some tests that exercise newer and more abstract C++ features like <algorithms> and <ranges> (#7)
* Added some examples to test compatibility with <algorithm> and ranges

* Fix up the ranges test a bit
Turns out it worked the first time, the compiler headers were just broken.
The resulting error noise hid the fact that it was trying to pass
an rfloat to std::round(), which fails. Clearing both of those
fixed the issue. Ranges work nicely.

* Remove the ENABLE_RANGE define
clang-14/15 support for ranges is very broken, and it's still the default
on the CI pipelines.