This change plays around with the idea of moving everything under rstd I'm not a big fan of the existing rmath namespace. It's just different enough from rcmath that I accidentally type 'rcmath::' all the time, but rcmath doesn't adhere to the one letter conversion rule I've been trying to follow.
It'd be nice to provide std overloads, but that's UB for obvious reasons.

It's also possible to simply have the wrappers as friend functions
and use ADL to select them, but this is only compatible with the c free
function declaration style. It's also a lot more noise in the header.

This still leaves us with an rmath namespace for the details stuff,
but it should be reasonable to move all of that under rstd in the next
commit.