Move all of the rcmath wrappers into <rfloat> header (#9)
* Move all of the rcmath wrappers into <rfloat> header
Unfortunately, we still can't use std, but this is at least not another
namespace.

* This change plays around with the idea of moving everything under rstd
I'm not a big fan of the existing rmath namespace. It's just different
enough from rcmath that I accidentally type 'rcmath::' all the time,
but rcmath doesn't adhere to the one letter conversion rule
I've been trying to follow.

It'd be nice to provide std overloads, but that's UB for obvious reasons.

It's also possible to simply have the wrappers as friend functions
and use ADL to select them, but this is only compatible with the c free
function declaration style. It's also a lot more noise in the header.

This still leaves us with an rmath namespace for the details stuff,
but it should be reasonable to move all of that under rstd in the next
commit.

* Remove all direct accesses from the library to mitigate Hyrum's law
It's possible that certain use cases built on top of this library
might not be able to uphold the same guarantees about the implementation
being solely a wrapper around a value of the underlying type.
Plus, it's more sanitary to have an explicit declaration of "I want to
take out the data member".

* Move everything into the rstd namespace
This cleans up the rmath/rstd distinction, and also hides some of the
declarations so they don't collide with user code.

* Update the docs and macros to reflect rmath going away