Initial testing step for adding QEMU CI runners for testing other platforms (#13)
* Add a QEMU CI pipeline to test reproducibility across a variety of platforms. 

QEMU tests need very different steps for building and running, so it
doesn't make sense to have them in the normal pipeline. This also reduces the
combinatorial explosion in the build matrix.

As a consequence of this, we need to build the QEMU tests with a header-only test framework.
GTest requires a binary library linked in, which is annoying, but the tests don't really need all of
the features GTest provides. nstead,
try a header only test library to see if the QEMU tests will run
and evaluate switching over since this library doesn't use the fancier
GTest features that require the binary.

For the sake of using the same code everywhere, every test had to be moved over to the new test library
on all platforms, and all tests required minor updates as a result.