Implement clang builds for emulated tests  (#16)
Introduce Clang++ to QEMU pipelines. This should have been relatively straightforward, but it uncovered a reproducibility issue in Clang itself that required introducing a new set of "compiler bug tests".

The issue is that when -ffast-math is passed as a flag, clang's sqrt() implementation becomes non-reproducible, but only on the ppc64le platform and only in highly specific conditions with specific inputs. The failing tests are disabled for the QEMU pipeline of clang + ppc64le, but remain enabled for all other pipelines.