Optimize clang and GCC codegen with compiler barrier intrinsics (#27)
* Try optimizing rfloat with compiler barrier intrisincs

This makes the barrier intrinsics the default implementation now for most
configurations and locks the inline assembly flags behind BARRIER_IMPL_ASM

The exception is Clang/LLVM, which has a broken barrier intrinsic. The
issue seems to be that clang's backend doesn't respect the contraction
metadata on IR emitted by the frontend and will occasionally
still produce fma() instructions in certain situations. This means that
any usage of fp-contract=off and protect-parens is probably broken.
Unbelievable.

We can work around this by taking advantage of the barrier intrinsic's
better optimization to do the work and insert a noop assembly block that
doesn't constrain the generated assembly. This has some minor overhead,
but it's relatively limited. We could reduce it further if the fp-contract
flags were available in the preprocessor, but they're not.
Clang with -ffast-math enabled still has to drop back to the inline
assembly hack and pay the full costs associated with it though.

This commit also fixes a silly error where Clang builds were excluding
sqrt() when they didn't need to. Oops

* Remove unnecessary SAFE_BINOP calls to improve optimization

* Update random compiler test pipeline and readme
The random flags pipeline is updated with the potentially problematic
flags to ensure they receive testing coverage

The readme has been rewritten to reflect the new compromises.